Updated handling for auto off handling of turnouts so it's now as described in Z21 protocol 5.2 LAN_X_SET_TURNOUT. Also added handling if auto off is not used (by default it is on).
Turnouts (decoders) general have a automatic switch off to avoid burned out coils. (Very) old turnout decoders may not the option to disable the output ,and also it is not assured all command stations send an off command after a turnout is set. So the WMC sends an off command after 500 msec....

Because of the added option the EEPROM configuration is changed and after updating a default WMC will appear...

Auto off handling needs to be tested!!!